@adamchainz 2020-09-21
> I wonder if there's anyway we could prevent regressions in this area in the future

Same. It can be hard to write automated tests.

I'm trying a few more tests right now and notcie that a lot of import time is actually spent compiling regexes in standard library modules. We moved most of our regexes to be lazy in e3d0b4d5501c6d0bc39f035e4345e5bdfde12e41 - there's probably an argument to allow the same in the actual `re` module.