@adamchainz 2020-09-21
Aha... closed issue because it's so hard to convince every dependency to update 😅 https://bugs.python.org/issue31580