@felixxm 2020-09-22
There is a small change here, previously we raised `ImportError` on a `backend` initialization (`self.backend({...})`) now we raises `ModuleNotFoundError` on access. I think it's acceptable.