@pope1ni 2020-09-22
> I think it's acceptable.

Agreed. I can't see why you'd access `backend` and not intend to use it.

Also this feels much cleaner than raising `ImportError` manually, especially now we get `ModuleNotFoundError`.